+++ CHANGELOG.md
@@ -5,9 +5,15 @@ All notable changes to this project will be documented in this file.
 ## [0.4.4] UNRELEASED
 ### Added
 - `Evaluate` class now implements a `__hash__` and `tolist()` methods.
+- BLEU 1 to n classes and functions.
 
 ### Changed
 - Function `get_install_info` now returns `package_path`.
+- AACMetric now indicate the output type when using `__call__` method.
+- Rename `AACEvaluate` to `DCASE2023Evaluate` and use `dcase2023` metric set instead of `all` metric set.
+
+### Fixed
+- `sbert_sim` name in internal instantiation functions.
 
 ## [0.4.3] 2023-06-15
 ### Changed